style: format Quiz
