+++ src/pages/Quiz/index.jsx
@@ -1,13 +1,8 @@
-// utils
-import { setInDataBase } from "../../utils/set-in-database.util";
-
 // hooks
 import { useEffect, useState } from "react";
 import { useRandom } from "../../hooks/random.hook";
 import { useSelector, useDispatch } from "react-redux";
 import { useHistory } from "react-router";
-import { memo } from "react";
-
 // actions
 import {
   decreaseScoreAction,
@@ -18,22 +13,22 @@ import {
   startLoadingAction,
   setFastestAnswerAction,
 } from "../../store/quiz/quiz.actions";
-
 // components
 import { Question } from "./components/Question";
 import { If } from "../../components/If";
 import { Else } from "../../components/Else";
-
-// global components
 import { LoadingSpinner } from "../../components/LoadingSpinner";
+// utils
+import { setInDataBase } from "../../utils/set-in-database.util";
 import {
   blankDataCreator,
   standardDataCreator,
   storeRightAnswerData,
   storeUnansweredData,
   storeWrongAnswerData,
-} from "./utils/store-extradata.util";
+} from "./utils/store-extra-data.util";
 
+import { memo } from "react";
 const QuestionMemo = memo(Question);
 
 // we use this function to get value from the store we write it here for sake of optimization
